refactor: move health/status/config to routes/general.py, delete litestar_app shim
